GlobalExceptionMiddleware created, CustomResponseDto created

+++ BookStore.Service/Concrete/UserService.cs
@@ -3,6 +3,7 @@ using BookStore.Core.Abstracts.Repositories;
 using BookStore.Core.Abstracts.Services;
 using BookStore.Core.Entities;
 using BookStore.Core.UnitOfWorks;
+using BookStore.Service.Exceptions;
 using BookStoreAPI.DTO.User.Request;
 using BookStoreAPI.DTO.User.Response;
 
@@ -13,7 +14,7 @@ namespace BookStore.Service.Concrete
         private readonly IUserRepository _userRepository;
         private readonly IMapper _mapper;
         private readonly IUnitOfWork _unitOfWork;
-        public UserService(IUserRepository userRepository, IMapper mapper,IUnitOfWork unitOfWork)
+        public UserService(IUserRepository userRepository, IMapper mapper, IUnitOfWork unitOfWork)
         {
             _userRepository = userRepository;
             _mapper = mapper;
@@ -23,6 +24,8 @@ namespace BookStore.Service.Concrete
         public async Task ActivateUserAsync(int id)
         {
             var user = await _userRepository.ActivateUserByIdAsync(id);
+            if (user == null)
+                throw new NotFoundException($"UserId: {id} not found.");
             _userRepository.UpdateAsync(user);
             await _unitOfWork.CommitAsync();
         }
@@ -30,6 +33,8 @@ namespace BookStore.Service.Concrete
         public async Task DeactivateUserAsync(int id)
         {
             var user = await _userRepository.DeactivateUserByIdAsync(id);
+            if (user == null)
+                throw new NotFoundException($"UserId: {id} not found.");
             _userRepository.UpdateAsync(user);
             await _unitOfWork.CommitAsync();
         }
@@ -37,6 +42,8 @@ namespace BookStore.Service.Concrete
         public async Task<GetUserDto> FindUserByIdAsync(int id)
         {
             var user = await _userRepository.FindUserByIdAsync(id);
+            if (user == null)
+                throw new NotFoundException($"UserId: {id} not found.");
             var userDto = _mapper.Map<GetUserDto>(user);
             return userDto;
         }
@@ -56,12 +63,16 @@ namespace BookStore.Service.Concrete
         public async Task MakeAdminUserAsync(string username)
         {
             var user = await _userRepository.MakeAdminAsync(username);
+            if (user == null)
+                throw new NotFoundException($"Username: {username} not found.");
             _userRepository.UpdateAsync(user);
             await _unitOfWork.CommitAsync();
         }
         public async Task SoftDeleteUserAsync(int id)
         {
             var user = await _userRepository.GetByIdAsync(id);
+            if (user == null)
+                throw new NotFoundException($"UserId: {id} not found.");
             _userRepository.Remove(user);
             await _unitOfWork.CommitAsync();
         }
@@ -69,16 +80,17 @@ namespace BookStore.Service.Concrete
         public async Task<GetUserDto> UpdateUserAsync(int id, UpdateUserDto updateUser)
         {
             var user = await _userRepository.FindUserByIdAsync(id);
-            if(user != null)
-            {
-                _mapper.Map(updateUser,user);
-                _userRepository.UpdateAsync(user);
-                await _unitOfWork.CommitAsync();
-                var userDto = _mapper.Map<GetUserDto>(user);
-                return userDto;
-            }
-            return null;
-            // Exception will be added.
+            if (user == null)
+                throw new NotFoundException($"UserId: {id} not found.");
+
+            _mapper.Map(updateUser, user);
+            _userRepository.UpdateAsync(user);
+            await _unitOfWork.CommitAsync();
+            var userDto = _mapper.Map<GetUserDto>(user);
+            return userDto;
+
+
+
         }
 
         //public async Task<IEnumerable<UsersWithBooksDto>> GetUsersWithBooksAsync()
